pyproject.toml
[project]
name = "chords"
version = "0.1.0"
description = "Accessible chord reference app with audio playback"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "prismatoid>=0.9.0",
    "pygame>=2.6.1",
    "sounddevice>=0.5.5",
]

[project.scripts]
chords = "main:main"

[dependency-groups]
dev = [
    "nuitka>=4.0.7",
]

[tool.nuitka]
standalone = true
onefile = true
windows-console-mode = "force"
output-filename = "chords"
include-data-files = []
